backend/requirements.txt
# API and Web Framework
fastapi>=0.110.0
uvicorn>=0.27.1
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.27
alembic>=1.13.1
psycopg2-binary>=2.9.9

# AWS
boto3>=1.34.34

# Celery
celery>=5.4.0
redis>=5.0.1

# Settings and Environment
pydantic>=2.6.1
pydantic-settings>=2.1.0
python-dotenv>=1.0.1

# Testing
pytest>=8.0.0
httpx>=0.27.0
moto>=5.0.2
pytest-asyncio>=0.23.5

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai>=1.12.0 